CAUSEWAY-4006: factors out NumericValueSemantics
as a base for all numerical values
