+++ api/applib/src/main/java/org/apache/causeway/applib/value/semantics/ValueSemanticsAbstract.java
@@ -18,13 +18,6 @@
  */
 package org.apache.causeway.applib.value.semantics;
 
-import java.math.BigDecimal;
-import java.math.BigInteger;
-import java.text.DecimalFormat;
-import java.text.DecimalFormatSymbols;
-import java.text.NumberFormat;
-import java.text.ParseException;
-import java.text.ParsePosition;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeFormatterBuilder;
 import java.time.format.FormatStyle;
@@ -32,7 +25,6 @@ import java.util.Locale;
 import java.util.Optional;
 import java.util.function.Function;
 import java.util.function.Supplier;
-import java.util.function.UnaryOperator;
 
 import jakarta.inject.Provider;
 
@@ -42,13 +34,11 @@ import org.jspecify.annotations.Nullable;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import org.apache.causeway.applib.annotation.TimePrecision;
-import org.apache.causeway.applib.exceptions.recoverable.TextEntryParseException;
 import org.apache.causeway.applib.fa.FontAwesomeLayers;
 import org.apache.causeway.applib.locale.UserLocale;
 import org.apache.causeway.applib.services.bookmark.IdStringifier;
 import org.apache.causeway.applib.services.i18n.TranslationContext;
 import org.apache.causeway.applib.services.i18n.TranslationService;
-import org.apache.causeway.applib.services.iactn.InteractionContext;
 import org.apache.causeway.applib.services.render.PlaceholderRenderService;
 import org.apache.causeway.applib.services.render.PlaceholderRenderService.PlaceholderLiteral;
 import org.apache.causeway.applib.util.schema.CommonDtoUtils;
@@ -115,10 +105,7 @@ ValueSemanticsProvider<T> {
      * @return {@link Locale} from given context or else system's default
      */
     protected UserLocale getUserLocale(final ValueSemanticsProvider.@Nullable Context context) {
-        return Optional.ofNullable(context)
-                .map(ValueSemanticsProvider.Context::interactionContext)
-                .map(InteractionContext::getLocale)
-                .orElseGet(UserLocale::getDefault);
+        return ValueSemanticsProvider.getUserLocale(context);
     }
 
     protected String renderTitle(final T value, final Function<T, String> toString) {
@@ -132,14 +119,6 @@ ValueSemanticsProvider<T> {
                 .map(toString)
                 .orElseGet(()->getPlaceholderRenderService().asHtml(PlaceholderLiteral.NULL_REPRESENTATION));
     }
-
-    protected String renderHtml(final T value, final Function<T, String> toString, final UnaryOperator<String> htmlPostprocessor) {
-        return Optional.ofNullable(value)
-                .map(toString)
-                .map(htmlPostprocessor)
-                .orElseGet(()->getPlaceholderRenderService().asHtml(PlaceholderLiteral.NULL_REPRESENTATION));
-    }
-
 
     // -- COMPOSITION UTILS
 
@@ -199,96 +178,30 @@ ValueSemanticsProvider<T> {
 
     // -- NUMBER FORMATTING/PARSING
 
-    /**
-     * @param context - nullable in support of JUnit testing
-     * @return {@link NumberFormat} the default from given context's locale
-     * or else system's default locale
-     *
-     * @implNote the format's MaximumFractionDigits are initialized to 16, as
-     * 64 bit IEEE 754 double has 15 decimal digits of precision;
-     * this is typically overruled later by implementations of
-     * {@link #configureDecimalFormat(org.apache.causeway.applib.adapters.ValueSemanticsProvider.Context, DecimalFormat) configureDecimalFormat}
-     */
-    protected DecimalFormat getNumberFormat(
-            final ValueSemanticsProvider.@Nullable Context context) {
-        return getNumberFormat(context, FormatUsageFor.RENDERING);
-    }
-
-    protected DecimalFormat getNumberFormat(
-            final ValueSemanticsProvider.@Nullable Context context,
-            final @NonNull FormatUsageFor usedFor) {
-        var format = (DecimalFormat)NumberFormat.getNumberInstance(getUserLocale(context).numberFormatLocale());
-        // prime w/ 16 (64 bit IEEE 754 double has 15 decimal digits of precision)
-        format.setMaximumFractionDigits(16);
-        configureDecimalFormat(context, format, usedFor);
-        return format;
-    }
-
-    protected Optional<BigInteger> parseInteger(
-            final ValueSemanticsProvider.@Nullable Context context,
-            final @Nullable String text) {
-        var input = _Strings.blankToNullOrTrim(text);
-        if(input==null) {
-			return Optional.empty();
-		}
-        try {
-            return parseDecimal(context, input, GroupingSeparatorPolicy.ALLOW)
-                    .map(BigDecimal::toBigIntegerExact);
-        } catch (final NumberFormatException | ArithmeticException e) {
-            throw new TextEntryParseException("Not an integer value " + text, e);
-        }
-    }
-
-    protected enum GroupingSeparatorPolicy {
-        ALLOW,
-        DISALLOW,
-        ;
-    }
-
-    protected Optional<BigDecimal> parseDecimal(
-            final @Nullable Context context,
-            final @Nullable String text,
-            final GroupingSeparatorPolicy groupingSeparatorPolicy) {
-        var input = _Strings.blankToNullOrTrim(text);
-        if(input==null) {
-			return Optional.empty();
-		}
-
-        if (groupingSeparatorPolicy == GroupingSeparatorPolicy.DISALLOW) {
-            var userLocale = getUserLocale(context);
-            var decimalFormatSymbols = new DecimalFormatSymbols(userLocale.numberFormatLocale());
-            var groupingSeparatorChar = decimalFormatSymbols.getGroupingSeparator();
-            if (input.contains(""+groupingSeparatorChar)) {
-				throw new TextEntryParseException("Invalid value '" + input + "'; do not use the '" + groupingSeparatorChar + "' grouping separator");
-			}
-        }
-
-        var format = getNumberFormat(context, FormatUsageFor.PARSING);
-        format.setParseBigDecimal(true);
-
-        var position = new ParsePosition(0);
-        try {
-            var number = (BigDecimal)format.parse(input, position);
-            if (position.getErrorIndex() != -1) {
-				throw new ParseException("could not parse input='" + input + "'", position.getErrorIndex());
-			} else if (position.getIndex() < input.length()) {
-				throw new ParseException("input='" + input + "' was not processed completely", position.getIndex());
-			}
-            // check for maxFractionDigits if required ...
-            final int maxFractionDigits = format.getMaximumFractionDigits();
-            if(maxFractionDigits>-1
-                    && number.scale()>format.getMaximumFractionDigits()) {
-				throw new TextEntryParseException(String.format(
-                        "No more than %d digits can be entered after the decimal separator, "
-                                + "got %d in '%s'.", maxFractionDigits, number.scale(), input));
-			}
-            return Optional.of(number);
-        } catch (final NumberFormatException | ParseException e) {
-            throw new TextEntryParseException(String.format(
-                    "Not a decimal value '%s': %s", input, e.getMessage()),
-                    e);
-        }
-    }
+//    /**
+//     * @param context - nullable in support of JUnit testing
+//     * @return {@link NumberFormat} the default from given context's locale
+//     * or else system's default locale
+//     *
+//     * @implNote the format's MaximumFractionDigits are initialized to 16, as
+//     * 64 bit IEEE 754 double has 15 decimal digits of precision;
+//     * this is typically overruled later by implementations of
+//     * {@link #configureDecimalFormat(org.apache.causeway.applib.adapters.ValueSemanticsProvider.Context, DecimalFormat) configureDecimalFormat}
+//     */
+//    protected DecimalFormat getNumberFormat(
+//            final ValueSemanticsProvider.@Nullable Context context) {
+//        return getNumberFormat(context, FormatUsageFor.RENDERING);
+//    }
+//
+//    protected DecimalFormat getNumberFormat(
+//            final ValueSemanticsProvider.@Nullable Context context,
+//            final @NonNull FormatUsageFor usedFor) {
+//        var format = (DecimalFormat)NumberFormat.getNumberInstance(getUserLocale(context).numberFormatLocale());
+//        // prime w/ 16 (64 bit IEEE 754 double has 15 decimal digits of precision)
+//        format.setMaximumFractionDigits(16);
+//        configureDecimalFormat(context, format, usedFor);
+//        return format;
+//    }
 
     public static enum FormatUsageFor {
         PARSING,
@@ -296,12 +209,6 @@ ValueSemanticsProvider<T> {
         public boolean isParsing() { return this==PARSING; }
         public boolean isRendering() { return this==RENDERING; }
     }
-
-    /**
-     * Typically overridden by BigDecimalValueSemantics to set min/max fractional digits.
-     */
-    protected void configureDecimalFormat(
-            final Context context, final DecimalFormat format, final FormatUsageFor usedFor) {}
 
     // -- TEMPORAL RENDERING
 
@@ -412,7 +319,5 @@ ValueSemanticsProvider<T> {
         return """
             <span>%s%s</span>""".formatted(faLayers.toHtml(), titleHtml);
     }
-    
-    
 
 }
