+++ core/metamodel/src/main/java/org/apache/causeway/core/metamodel/valuesemantics/BigIntegerValueSemantics.java
@@ -31,6 +31,7 @@ import org.springframework.stereotype.Component;
 import org.apache.causeway.applib.annotation.PriorityPrecedence;
 import org.apache.causeway.applib.services.bookmark.IdStringifier;
 import org.apache.causeway.applib.value.semantics.DefaultsProvider;
+import org.apache.causeway.applib.value.semantics.NumericValueSemantics;
 import org.apache.causeway.applib.value.semantics.Parser;
 import org.apache.causeway.applib.value.semantics.Renderer;
 import org.apache.causeway.applib.value.semantics.ValueDecomposition;
@@ -45,6 +46,7 @@ import org.apache.causeway.schema.common.v2.ValueWithTypeDto;
 public class BigIntegerValueSemantics
 extends ValueSemanticsAbstract<BigInteger>
 implements
+    NumericValueSemantics,
     DefaultsProvider<BigInteger>,
     Parser<BigInteger>,
     Renderer<BigInteger>,
@@ -87,7 +89,7 @@ implements
 
     @Override
     public String htmlPresentation(final Context context, final BigInteger value) {
-        return renderHtml(value, getNumberFormat(context)::format, super::toMonospace);
+        return renderHtml(value, pipe(getNumberFormat(context)::format, super::toMonospace));
     }
 
     // -- ID STRINGIFIER
@@ -114,7 +116,7 @@ implements
 
     @Override
     public BigInteger parseTextRepresentation(final Context context, final String text) {
-        return super.parseInteger(context, text)
+        return parseInteger(context, text)
                 .orElse(null);
     }
